backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=12.0

# Google Cloud & AI
google-adk>=1.5.0
google-generativeai>=0.8.3
annotateai>=0.1.0

# Firebase
firebase-admin>=6.5.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
